# swagger-spec.json
{"openapi":"3.0.0","paths":{"/users":{"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/users/{id}":{"post":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/recipes":{"post":{"operationId":"RecipesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRecipeDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"RecipesController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/recipes/{id}":{"post":{"operationId":"RecipesController_update","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRecipeDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"RecipesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"Yes!Chef API","description":"API description for Yes!Chef","version":"1.0","contact":{}},"tags":[{"name":"Yes!Chef","description":""}],"servers":[],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{"fullname":{"type":"string"},"gender":{"type":"string"},"email":{"type":"string"},"location":{"type":"string"},"dietary":{"type":"array","items":{"type":"string"}},"favorites":{"type":"array","items":{"type":"string"}}},"required":["fullname","gender","email","location","dietary","favorites"]},"UpdateUserDto":{"type":"object","properties":{"fullname":{"type":"string"},"gender":{"type":"string"},"email":{"type":"string"},"location":{"type":"string"},"dietary":{"type":"array","items":{"type":"string"}},"favorites":{"type":"array","items":{"type":"string"}}}},"CreateRecipeDto":{"type":"object","properties":{"name":{"type":"string"},"desc":{"type":"string"},"star":{"type":"number"},"time":{"type":"number"},"calo":{"type":"number"},"ingredients":{"type":"array","items":{"type":"string"}},"steps":{"type":"array","items":{"type":"string"}}},"required":["name","desc","star","time","calo","ingredients","steps"]},"UpdateRecipeDto":{"type":"object","properties":{"name":{"type":"string"},"desc":{"type":"string"},"star":{"type":"number"},"time":{"type":"number"},"calo":{"type":"number"},"ingredients":{"type":"array","items":{"type":"string"}},"steps":{"type":"array","items":{"type":"string"}}}}}}}